requirements.txt
asgiref==3.3.1
Django~=3.1
Pillow==8.1.0
python-decouple==3.3
pytz==2020.5
sqlparse==0.4.1
djangorestframework~=3.12.2
PyJWT==1.7.1
djangorestframework_simplejwt==4.5.0
django-rest-swagger==2.2.0
django-filter